Add setup of callbacks to notify updates of the client state
In the future, a management interface to support server migration is going to be implemented.
To update the state retrievable using the interface, it is necessary to add some callbacks
that are invoked when an event of interest happens in the server. This commit introduces
the setup (but not the invocation) of callbacks for the completion of a handshake,
the detection of a client migration and the end of a connection.
Even if the management interface is going to be developed as the last feature,
these callbacks are useful to implement some upcoming integration tests.