Store original server connection ID in the migration state
Before this commit, the migration/state callbacks were using directly the server CID stored in the serverConnectionId field inside QuicConnectionStateBase. Since serverConnectionId could change due to the reception of PATH_CHALLENGE or RETIRE_CONNECTION_ID, it is better to save a copy of the original CID separately in the migration state. This copy never changes and is going to be useful to identify correctly the transport when notifying an imminent server migration.