Change trigger for migration probing from packet marked loss to PTO
Previously, server migration probing was (attempted to be) started/updated by manually scheduling PING frames when a packet was marked as lost. This approach is wrong, because QUIC employs the PTO mechanism before declaring a packet as lost: after the server changes address, packets sent by the client to the old server address are not marked as lost (thus retransmitted) until ack-eliciting packets sent by the client itself using the PTO mechanism are acknowledged by the server. Then, since no calls to markPacketLoss() and consequently to maybeStartServerMigrationProbing() were done, the peer address of the client was never updated, the migration probing was never detected, and the PTO-generated probes never acknowledged, leading to an "Exceeded max PTO" error. With this commit, the server migration probing detection is bound to onPTOAlarm() invocations, solving the issue.