Change default active connection id limit
The number of CIDs issued after the handshake is equal to the minimum between the limit advertised by the peer and the default value identified by kDefaultActiveConnectionIdLimit. If the latter is too small, it is impossible to migrate a client/server multiple times (and setting selfActiveConnectionIdLimit in the transport settings becomes useless).
Originally, kDefaultActiveConnectionIdLimit was equal to 2; then, it was set to 10 after introducing server migration. Now, this limit is increased to 100, giving more freedom to tune selfActiveConnectionIdLimit.
Given that CIDs are not sent by an endpoint if the migration of the peer is disabled, this change does not affect the performance of normal connections (migration is disabled by default).